+++ packages/base/.storybook/main.ts
@@ -15,6 +15,8 @@ const config: StorybookConfig = {
     getAbsolutePath("@storybook/addon-links"),
     getAbsolutePath("@storybook/addon-essentials"),
     getAbsolutePath("@storybook/addon-interactions"),
+    getAbsolutePath("@storybook/addon-a11y"),
+    "storybook-addon-performance"
   ],
   framework: {
     name: getAbsolutePath("@storybook/react-vite"),