+++ packages/base/src/components/Buttons/Button/Button.styled.tsx
@@ -70,6 +70,7 @@ export const StyledButton = styled.button<ButtonOptionProps>`
   display: flex;
   align-items: center;
   justify-content: center;
+  height: fit-content;
   gap: 4px;
   padding: 0 12px;
   border-radius: 6px;