+++ packages/base/src/components/Inputs/Input/Input.styled.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { colorTokens } from '../../../styles';
 
 export const StyledInput = styled.input`
-  padding: 4px 8px 4px 12px;
+  padding: 4px 0 4px 12px;
   background-color: transparent;
   width: 100%;
   border: none;
@@ -16,4 +16,5 @@ export const RightElementWrapper = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
+  padding: 0 8px 0 0;
 `;