+++ packages/base/src/components/Inputs/TextArea/TextArea.tsx
@@ -21,6 +21,7 @@ const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
       onBlur,
       onFocus,
       minHeight = 100,
+      className,
       ...rest
     },
     ref,
@@ -82,7 +83,7 @@ const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
     return (
       <InputWrapper
         htmlFor={uniqueId}
-        focused={isFocused}
+        isFocused={isFocused}
         disabled={disabled}
         isHovered={isHovered}
         isError={isError}
@@ -92,7 +93,7 @@ const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
         data-shoplflow={'text-area'}
       >
         <StyledTextarea
-          className='body1_400'
+          className={'body1_400' + (className ? ` ${className}` : '')}
           placeholder={placeholder}
           onFocus={handleOnFocus}
           onBlur={handleOnBlur}