+++ packages/base/src/components/Popper/Popper.styled.ts
@@ -1,8 +1,9 @@
 import styled from '@emotion/styled';
+import type { PopperOptionProps } from './Popper.types';
 
-export const StyledPopper = styled.div`
+export const StyledPopper = styled.div<Pick<PopperOptionProps, 'height' | 'width'>>`
   display: flex;
   flex-direction: row;
-  width: 100%;
-  height: 100%;
+  width: ${({ width }) => width ?? 'fit-content'};
+  height: ${({ height }) => height ?? 'fit-content'};
 `;