+++ packages/base/src/components/index.ts
@@ -7,6 +7,7 @@ export * from './Switch';
 export * from './Chips';
 export * from './Buttons';
 export * from './Callout';
+export * from './Popper';
 export * from './Icon';
 export * from './Inputs';
 export * from './ControlButtons';