+++ packages/base/src/hooks/index.ts
@@ -3,4 +3,3 @@ export * from './useDomain';
 // 모달 상태 관리
 export * from './useHandleModal';
 export * from './useModalValue';
-// export * from './useModalStore';