+++ packages/base/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @shoplflow/base
 
+## 0.12.4
+
+### Patch Changes
+
+- [#113](https://github.com/shopl/shoplflow/pull/113) [`8ed72c2`](https://github.com/shopl/shoplflow/commit/8ed72c245fabc1ad4ad9cfe0872e67943c692d68) Thanks [@bong9boy](https://github.com/bong9boy)! - 400->font weight medium으로 일괄 변경
+
+## 0.12.3
+
+### Patch Changes
+
+- Updated dependencies [[`668681a`](https://github.com/shopl/shoplflow/commit/668681afdc78faf45e9654df9f382b779726c964)]:
+  - @shoplflow/hada-assets@0.0.3
+  - @shoplflow/shopl-assets@0.3.1
+
+## 0.12.2
+
+### Patch Changes
+
+- [#109](https://github.com/shopl/shoplflow/pull/109) [`ee6ea4b`](https://github.com/shopl/shoplflow/commit/ee6ea4b7dc1870fd88e764ca9e6cfe298cb76f46) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - iconButton sizeVar, styleVar 타입 추론이 안되는 이슈 수정
+
+- [#109](https://github.com/shopl/shoplflow/pull/109) [`ee6ea4b`](https://github.com/shopl/shoplflow/commit/ee6ea4b7dc1870fd88e764ca9e6cfe298cb76f46) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - input 좌측 padding이 올바르게 적용되지 않는 문제 수정
+
+## 0.12.1
+
+### Patch Changes
+
+- [#107](https://github.com/shopl/shoplflow/pull/107) [`61228e5`](https://github.com/shopl/shoplflow/commit/61228e53d42a1077b46995edf9bc3f1b6036b518) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - package.json type 변경
+
+- Updated dependencies [[`61228e5`](https://github.com/shopl/shoplflow/commit/61228e53d42a1077b46995edf9bc3f1b6036b518)]:
+  - @shoplflow/utils@0.3.1
+
+## 0.12.0
+
+### Minor Changes
+
+- [#103](https://github.com/shopl/shoplflow/pull/103) [`2c8286e`](https://github.com/shopl/shoplflow/commit/2c8286eef69dec291d088c783072cc80d4d9ddee) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - 스토리 북의 페이지를 이동해도 도메인이 유지됩니다.
+
+- [#103](https://github.com/shopl/shoplflow/pull/103) [`2c8286e`](https://github.com/shopl/shoplflow/commit/2c8286eef69dec291d088c783072cc80d4d9ddee) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - input 컴포넌트 추가
+
+- [#103](https://github.com/shopl/shoplflow/pull/103) [`2c8286e`](https://github.com/shopl/shoplflow/commit/2c8286eef69dec291d088c783072cc80d4d9ddee) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - modal 아키텍쳐 변경 zustand -> react context
+
 ## 0.11.4
 
 ### Patch Changes