+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shoplflow/utils
 
+## 0.3.1
+
+### Patch Changes
+
+- [#107](https://github.com/shopl/shoplflow/pull/107) [`61228e5`](https://github.com/shopl/shoplflow/commit/61228e53d42a1077b46995edf9bc3f1b6036b518) Thanks [@Jason-Jeong-Dev](https://github.com/Jason-Jeong-Dev)! - utils package.json type 수정
+
 ## 0.3.0
 
 ### Minor Changes